chatBot/gitasc/requirement.txt
sentencepiece==0.1.99
torch==2.1.2
torchvision==0.16.2

aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cmake==3.27.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.2
datasets==2.16.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
distro==1.8.0
dlib==19.24.2
dnspython==2.4.2
EasyProcess==1.1
email-validator==2.1.0.post1
emoji==2.8.0
entrypoint2==1.1
entrypoints==0.4
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.7.4
Faker==20.1.0
fastapi==0.104.1
favicon==0.7.0
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google_search_results==2.4.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.9
langchain-openai==0.0.2
langchainhub==0.1.14
langdetect==1.0.9
langsmith==0.0.79
llama-index==0.9.33
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.16.2
openai==1.7.1
opencv-python==4.8.1.78
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
posthog==3.0.2
protobuf==4.25.1
pulsar-client==3.3.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pymdown-extensions==10.5
PyMsgBox==1.0.9
pyparsing==3.1.1
pypdf==3.17.1
PyPDF2==3.0.1
pypdfium2==4.24.0
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
pytweening==1.0.7
pytz==2023.3
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.0
rapidfuzz==3.5.2
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
starlette==0.27.0
streamlit==1.28.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
text-generation==0.6.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.0
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.24.6
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.1
xxhash==3.4.1
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.17.0
